Fix install paths
Previously we were installing all our headers into `include`, instead of
preserving the directory structure in there.  This breaks lots of
things, for instance we create an `assert.h` that'll break things
attempting to include the stdlib `assert.h`.

This preserves directory structure correctly when running `cmake
--build --target install` (and therefore also `pip install .`,
which calls cmake).  Tested locally in a conda environment.