[SymForce-External] Workflow and build support for pip wheels
See `documentation.rst` for usage description.  This primarily consists
of a workflow (manually triggered) to build wheels on GitHub Actions,
and supporting changes to our build to make this work, i.e. for the
right files to end up in the wheels in the right places (including
getting the README to look right on PyPI).

You can now:
- Build a wheel in a fresh environment, with basically whatever
  method you'd like (`python -m build`, `pip wheel`,
  `python setup.py bdist_wheel`).  Running this repeatedly to build
  multiple wheels also works, but may require cleaning the build
  directory depending on the method used (see the TODO on symenginepy
  in CMakeLists.txt)
- `pip install .` repeatedly to re-install symforce in the same
  environment
- `pip install -e .` repeatedly, manually cleaning the build
  (technically just the symenginepy build) between runs

Closes symforce-org/symforce#157

GitOrigin-RevId: 0241cd20589e070c8e3d857f61fc2132746113c9