[SymForce] Call modify_symbolic_api once per sympy
Before this commit, we would not call `modify_symbolic_api`
if you set the backend to the current backend, but would call
it a second time if you, say, set the backend to sympy and then
back to symengine.

Since many of the functions called by `modify_symbolic_api` call
the sympy methods they are overriding (see:
- override_symbol_new
- override_subs
- override_solve
- add_custom_methods (atan2)),
calling them twice more or less applies the modifications twice.

For example, the scopes added to the `symforce.sympy.Symbol.__init__`
will be applied multiple times.

This fixes the issue by adding a flag to `symforce.sympy` to ensure
we only apply the modifications once.

Not currently unit tested because we can't change the backend in the
unit test, so would have to use a subprocess or something. Not doing
this right now because busy with other stuff.

Topic: modify_symbolic_api_only_once
GitOrigin-RevId: c51db1ed4c3f893bef46f90fc460d81fcf56fa23