[SymForce] Change format_py_dir to take T.Openable
This didn't actually require any work. Works just
fine with a `Path` and I'm sure will work well with
other `os.Pathlike`s.

Also, the only other path argument in the file is
filename of `format_cpp`, but it does seem that
it's most appropriate for it to be a `str`, so I
left that unchanged.

Topic: format_py_dir_arg_type
Relative: path_not_str
GitOrigin-RevId: 86c4f4476f63a8697ec32a0eb65e5f62ad2b1625