Rename sympy/symengine switching from "backend" to "symbolic API"
Because we are calling the codegen langauge targets "backends", it would be confusing to keep
doing so with the SymPy / SymEngine switch. Backend matches LLVM terminology, so it makes sense
there. We switch to using symbolic_api for the latter, which has some downsides in that we're
setting the implementation and not the API, but it seems like the best choice of the options
we've discussed.