[SymForce-External] Centralize code for creating new code generation backends
Move almost all backend-specific code into a sub-package of `symforce/codegen/backends`, rather than being scattered around multiple core files. This includes the code printer, codegen config, and a bunch of switch statements around the codegen machinery. The goal is to centralize and make it cleaner to add new codegen backends. These changes were made in parallel with adding a javascript backend, which will come in a follow-on review.

See this [README](https://github.com/symforce-org/symforce/blob/centralize-backend-specific-code/symforce/codegen/backends/README.md) for the steps to add a backend after this review.

There are still several ways to improve code quality and reorganize backend-specific code, but this is a step forward.

Closes symforce-org/symforce#158

GitOrigin-RevId: 0f97e336c7cac0da48173fcb69776a51470a5cc0