[SymForce] Use sm.epsilon() as default eps
Add sm.epsilon(). This is to make it easier
to change what value of epsilon is used everywhere.

It is a function as opposed to a constant for no particular
reason, but generally to be more flexible should the need
arise.

This will make it easier to change the default value of epsilon
for many functions with just a change to sm.epsilon().

Topic: use_default_epsilon
Relative: rename_default_to_numerical
GitOrigin-RevId: baf6557a12f75b9560510eec8debd1cdee6c965d