[SymForce-External] Add minimum spdlog version
Add minimum spdlog version

Set the minimum spdlog version to the minimum one that symforce currently successfully builds with.

This was tested by successfully building with all spdlog versions 1.9 and up. Everything below 1.9.0 failed.

I also updated the `GIT_TAG` to just use the tag name instead of the commit. I find it to be a little more readable, but maybe the commit hash is nice for other reasons.

Fixes symforce-org/symforce#182.

Closes symforce-org/symforce#184

GitOrigin-RevId: ab7586e7e94aa766efb3fac82993d53272f0d6d9