[SymForce] Make cc_sym types pickleable
Specifically, allow pickling of `Key`, `Linearization`, `OptimizationStats`,
and `Values`.

I do not make `Factor` and `Optimizer` pickleable because they contain
references to functions which cannot, in general, be pickled.

I also changed the return type of `OptimizationStats.best_linearization`
to a `Optional[cc_sym.Linearization]` (was just an `object` before). Had
to do it in a janky string replacement way because, while pybind11
supports optional types (and I imagine would generate the right return
type if we were to use them; but I haven't actually confirmed it), we
only use C++14, and std::optional isn't added until C++17.

I could have used the lcm type encoding to handle pickling of all the
types (but instead I only used it for `Values`), but I didn't because I
thought it would have made the pickling process seem more complicated
than it was/harder to understand. I did use it for `Values` because it
seemed to be the one case where doing so would be easier.

Fixes symforce-org/symforce#109

Topic: pickle_cc_sym
GitOrigin-RevId: 1751774550bc6b6351a6c2545969943f12e34f5c