[Patch] Preserve downloads symforce-org.github.io
It turns out there were some files and directories already present in
`symforce-org/symforce-org.github.io` which aren't regenerated by `make
docs` yet we still want to preserve.

Since redownloading the contents of `downloads` each time didn't seem
pleasant, and because I didn't want us to have to go through a big dance
every time we add a file to `symforce-org.github.io`, I am using a
modified version of the old action to push to the repo which has an
option to preserve certain files in the destination repo.