[SymForce] Remove local Catch2 (instead fetch it)
It's my understanding we only had a local copy because in the past it
wasn't possible to fetch from github.

This PR:
- Deletes the local copy of Catch2 and instead fetches it in our build
files
- Replaces the old mono-header with the split up headers. This is not
strictly necessary, as the new versions of Catch2 still have a
mono-header, but the authors of Catch2 recommend this change to speed up
compile times.
- Does not change the version of Catch2 that we use (v3.0.0-preview3)

We don't update to the latest version (v3.0.1) because it does not
compile on all the gcc compiler versions we'd like to support (for
example gcc-5).

As a note, one benefit of v3.0.1 over v3.0.0-preview3 is that v3.0.1
will compile on ubuntu-22.04 (see
https://github.com/symforce-org/symforce/issues/132)

**Overview of changes**

Needed to change the `CMakeLists.txt` to use the `Catch2::Catch2WithMain`
target (which we use instead of `Catch2::Catch2` because we don't define
a custom main method for any of our tests).

Deleted the old `third_party/catch2` because it's being replaced.

Most of the work was simply changing the includes to use the appropriate
headers instead of `#include "catch.hpp"`. Could have used `#include
<catch2/catch_all.hpp>`, but the catch2 docs recomends to not do this,
and instead include the specific headers to speed up compilation time
(at the cost of convenience)

Topic: update_catch2
GitOrigin-RevId: 5e16e149657a7807b8baf7209969a88bf4361dc0