[Symforce] Codegen ToYawPitchRoll for C++ Rot3
I changed `Rot3::ToYawPitchRoll()` to use the same implementation as we
do in the symbolic python version. Previously we were converting the
quaternion to a rotation matrix and then using built-in Eigen functions
to get the Euler angles, but this could give ambiguous results.

For example, if you construct a rotation with YPR (-10, 10, -10)
degrees, it can sometimes be reconstructed as something like YPR (170,
170, -180) degrees. Computing Euler angles directly from the quaternion
is much more stable (and it's probably much faster).

Topic: symforce_rot3_cpp_to_yaw_pitch_roll
GitOrigin-RevId: 85de520542ba265dced379a6655f42d536cd32d7