[SymForce] Make TemplateList.add arg order match
Specifically, make the argument order of `TemplateList.add` match the
argument order of `render_template`.

The order:
- `template_path` first because it is more or less the name of the
template, so it's easier to make sense of the call if it comes first
- data second because it is the only other argument that's always
required.
- template_dir next because, in the future, I'd like to make it a
required argument for `render_template`, but an optional argument for
`TemplateList.add`.
- Then `output_path` because it is optional for `render_template` and
there isn't a strong reason to not also have it be optional for
`TemplateList.add`.

In order to make `output_path` optional in `TemplateList.add`, I'm
mirroring the behavior of `render_template` by simply returning a list
of the rendered templates as strings.

Topic: render_template_arg_order
Relative: relative_render_template
GitOrigin-RevId: 71bee333681382c4c03d37a99c572264fcb1e654