Removed unused imports in symforce python function codegen template
I ran into an issue (bug?) where generating a python function that takes
and LCM type as an argument results in the namespace of the LCM type
being imported in the generated python file. However, we don't actually
use the import anywhere in the generated function (not even for typing
currently), so this just causes import errors if the package of the
generated LCM type is not on the python path.

Maybe we will want to add this back in some form in the future, but I
think we should remove this for now as it makes it annoying to import
generated functions which use LCM types.

Topic: symforce_no_namespace_import_in_py_codegen
GitOrigin-RevId: a1c955c30dc55d464c80dbf29825fbce27def902