Symforce Values.from_storage allow different scalar types
Noticed that when calling `from_storage` on a Values containing a
sequence of symbols of different types (e.g.
`symengine.lib.symengine_wrapper.Pow` and
`symengine.lib.symengine_wrapper.RealDouble`) the assertion that each
element in the sequence is the same type would be raised.

This was producing the behavior where we could call `evalf` on a
sequence of expresssions of different types, but could not call `evalf`
on a Values containing the same seqeunce.

Topic: symforce_values_from_storage_sequence_scalar_fix
GitOrigin-RevId: d30fc3eff5f67435fd0113baeebef743a9cfffb2