[SymForce] Don't add symengine to sys.modules if import fails
This was resulting in some weird behavior when the symengine module
exists but can't be imported successfully - we'll call
`_find_symengine`, which adds `symengine` to `sys.modules`, but then
will raise `ImportError` or something while executing the module.  If
this happened on `import symforce`, you could then call
`symforce.set_symbolic_api("symengine")` and `_find_symengine` would
take the first early return since `symengine` was still sitting in
`sys.modules` (just in an invalid state).  If we remove it instead like
we do here, you correctly get the `ImportError` when you call
`symforce.set_symbolic_api("symengine")`.  Confirmed this is the
behavior you get from normal `import` statements (failed imports result
in the module not being present in `sys.modules`).  Not sure why the
recommended snippet in the Python docs don't handle this (the version
without the try-except is copied verbatim from the docs).

Reviewers: hayk,bradley,nathan
Topic: sym-engine-safe-import
GitOrigin-RevId: 79ac82a6d00f4856818890bc24dfdb3cb6e9e45c