[SymForce] Reshape matrix inputs to be rank 2
This is to allow our generated code to take both rank 1 and rank 2
inputs.

Reshape to rank 2 and not rank 1 because reshaping a column major order
matrix to a rank 1 vector performs a copy (which we'd like to avoid).

To handle this, we index into the inputs as if they were rank 2 (which
they certainly are after reshaping), return vectors as rank 2 ndarrays
instead of lists, and type vectors as ndarrays.

Note, since now the generated `sym` classes (`Rot2`, `Rot3`, `Pose2`,
`Pose3`) take and return `ndarrays`, if they are to be placed in
symbolic `Values`, we need to make sure to convert the inputs and
outputs when they're methods are called by `Values` (this is done by
`SymClassLieGroupOps`.

Topic: reshape_ndarrays
Relative:
GitOrigin-RevId: 527e4985bf6a9b0b657b28344fbc4a08f07c8fa7