[SymForce] Codegen PoseX.position and rotation
Previously these were hand-generated. Want to codegen them instead so
that the indices to access aren't hardcoded, and it becomes easier for
changes to the symbolic code to propagate the generated code.

As a side note, some usages had to be updated to reflect the fact that
this now returns a 2d ndarray rather than a 1d ndarray.

In other words, this has a **BREAKING CHANGE**, as the shape of the
vectors returned by `PoseX.position` is now 2d (instead of 1d).

Also, some plumbing in util.jinja had to be changed to avoid calling
`sym.Rot3` even when `Rot3` had already been imported.

Topic: codegen_position_and_rotation
Relative:
GitOrigin-RevId: dcdbf631764b45d57ea1ae947633635ac0222ac2