[SymForce] Run python tests with ctest
This breaks coverage reports, but we haven't generated a new one of
those in months so I think it's ok to punt until we actually fix up
coverage and make it autogenerated on push to main and such.

Generally unittest and unittest-based tools have a pretty strong
assumption that it's fine to import all your tests into the same
interpreter.  Even e.g. `unittest-parallel` which can run all your tests
in their own subprocesses loads all of them into the parent process
(which means that when any of the tests are run in their subprocess, all
the tests have already been imported).  This is sort of a sign that
we're doing something bad by creating this kind of global state such
that this doesn't work, but I think we should be able to work around it.

Topic: ctest_python
Reviewers: hayk,bradley,aaron
Uploader:
GitOrigin-RevId: cfc49cf2320e3255746444867774596266b37cdf