[SymForce-External] Fix robot_2d_localization plotting
Fix robot_2d_localization plotting

Fixes symforce-org/symforce#238

This was broken by the python matrix shape changes.  It'd be nice if we
had test coverage on this, although that does seem potentially
difficult.

Also added setting epsilon to a symbol, which silences a bunch of
warnings.  Not sure if this setup is ideal.

Closes symforce-org/symforce#245

GitOrigin-RevId: c13f4936e8b9950f715c1cfa48fd7445e8930b49