sym class constructors accept column vectors
Previously, the class constructors would not accept column vectors.
This is inconvenient because our generated functions accept column
vectors and return column vectors.

So, this PR changes the constructors to accept column vectors as well
(more or less, as row vectors and other funky shapes are also accepted).

Also, replaced assertions on user inputs with exceptions, and asserted
that the user's input vectors, for example, aren't too long, as they
were previously able to silently create wonky Poses (for example.

Notably, this commit does not change the `from_storage` methods to
accept column vectors. This is because it's a bit less obvious that they
should accept column vectors, and I wanted to save worrying about it for
another day.

Fixes #246