Modify optimizer tutorial section of readme
In this section, the `Values` class was introduced and used for the
first time without a proper explanation. I thought this might be
confusing to new users and so have made these changes.

Also, provided a bit more forwarning on how exactly we'll optimize the
values, fixed a typo, and generally tried to improve the readability.