Update Catch2 to v3.1.1; Test on ubuntu jammy
This version of Catch2 doesn't have the same issues that sometimes
causes v3.0.1 to fail to compile with older ubuntu versions/compiler
versions.

Moreover, it also can compile with newer compilers (like gcc-11) on
ubuntu jammy.

So, adding ubuntu jammy to the `test_editable_pip_install` and `ci`
github action workflows.

To suport the Jammy build, I needed to make `setup.py` more robust to
different locations that the `symengine_wrapper.so` file might be
placed in after the build step.

Only real reason I'm not running gcc-5 on Jammy is because it is a bit
inconvenient to get a copy of it on Jammy, and it's not obviously worth
the hassle.

Set `PYTHON_EXECUTABLE` to equal `SYMFORCE_PYTHON` before adding
pybind11 to ensure that the whole build is referencing the same python
version (for example, when I had both python3.10 and python3.8 installed
on ubuntu jammy, `PYTHON_EXECUTABLE` was automatically set to python3.10
despite `SYMFORCE_PYTHON` being set to python3.8).

Fixes #132