[SymForce] Added assert to prevent codgen with nested databuffers
There exists a bug in codegen where the generated formatted symbols will
incorrectly be shifted when a databuffer is present in an input Values
object. E.g. when generating a function which takes an LCM object
containing a databuffer as an element, the generated function body will
not use the correct symbols for each scalar symbol after the databuffer
element in the LCM object storage order. Currently, the proper way to do
codegen with databuffers is to not include them in an input Values
object, but instead pass them as a separate top-level arg.

This adds an assert which prevents such symbol misalignment from
happening.

Topic: assert_no_nested_databuffers_codegen
GitOrigin-RevId: 64e9ca93cec005a2187a549be57f78b74089c32a