Require python LCM arguments to construct using `from_all_fields`
It's useful to be able to construct a python LCM type and know that you have supplied arguments for all fields.  It would be nice to make this the behavior of `__ini__()` (taking away default arguments there) and add a new `.default()` constructor that supplied all defaults, but that would change existing behavior and a lot of code.

Instead this defines the static constructor `from_all_fields()` with no default argument values that can be used to make sure you've supplied all fields.

Cannot call a function with more than 255 arguments until Python 3.7 or later, so disable generation of `from_all_fields` in this case: https://stackoverflow.com/questions/26766076/pass-more-than-255-arguments-to-a-function

Topic: peter/require_python_lcm_arguments
Relative:
Reviewers:
Assignees:
Branches:
GitOrigin-RevId: ceb1d2328d84a84ad75ba864a31a4b6ad4d15771