[ImuFactor] Add an on manifold Imu Factor
Unit testing is left to a follow up commit.

The core use-facing api is found in
- symforce/slam/imu_preintegration/preintegrated_imu_measurements.h
- symforce/slam/imu_preintegration/imu_factor.h

Unlike most other factors (/residual functions) in symforce, the
ImuFactor is a functor so as to store the pre-integrated values
with the function rather than in the `Values`. (There are a lot of
parameters that are constant and unique to each factor).

A python version of the `ImuFactor` class will come in a subsequent
commit.

Topic: on_manifold_imu_factor