Demo op count reduction if assume sqrt_info is ut
In many of our generated factors, we take the square root information
matrix as an argument, but make no assumptions about it's sparsity
pattern, assuming it's just a regular dense square matrix.

However, often times we pass in special matrices, for example diagonal
matrices or upper triangular matrices (which one might get if one uses
cholesky decomposition on the inverse of the covariance, for example).

As this commit demonstrates, passing this knowledge on to the symbolic
implementation of the code can reduce op counts considerably.

So, if we care about reducing op counts in these cases, a couple options
come to mind:
- Come up with a mechanism to specify that our sqrt_information matrices
have specific shapes so that this can be used during codegen
- Pull the sqrt_info matrix out of the residual and instead multiply the
outputs of the residual function after evaluation.

Topic: demo_sqrt_info_savings
Labels: draft