[SymForce] Add include_jacobians flag
To toggle whether the optimizer will allocate and fill out the problem
jacobian.  The jacobian is unnecessary for normal optimization, it's
only needed for 1) computing the linear error as a debug output or 2) if
you specifically want the jacobian for some other purpose.  This flag
defaults to false.

Note that each individual factor in the problem will still compute its
jacobian currently, it just won't be copied into the full problem
jacobian.

Topic: include-jacs
Relative: linearizer3
GitOrigin-RevId: 19b98678486fc51a1990f8b8544c2aae234bbdca