[SymForce] Allow setting default epsilon to same value
There's no reason not to allow this, and it's necessary for e.g. running
SymForce from something like streamlit that wants to call this
repeatedly when re-running a module that imports symforce (or you could
check if it's been set already, but that's sort of obnoxious and
nonobvious).

Reviewers: bradley,nathan,chao,hayk
Topic: sf-repeat-set-epsilon
GitOrigin-RevId: 933ac2b1547b4521e78941597115c5b3e845fd44