[SymForce] return 1d vectors if !return_2d_vectors
Change `python/templates/util/util.jinja` to return 1d ndarrays for
vector outputs in generated python code if `return_2d_vectors` is `False`.

The other idea was to instead return 1d ndarrays if `reshape_vectors`
was `True` (but not if `False`, because we wanted a vector output to
be a valid vector input). But this served to complicate the behavior,
potentially confusing people. It also doesn't make it easy for a user
who wants to return a MxN matrix as a 2 index object which just happens
to have one of M or N be 1.

Also, update `symforce_codegen_test.py` to expect 1d ndarrays in such
cases.

Topic: 1d_python_vector_outputs
GitOrigin-RevId: 8605b7466b27b1910a26a3e684d53c49dc231acc