[SymForce] Reduce the overload set of Factor constructors
Before we have a total of 12 functions related to Factor construction.
Because we have to overload for both versions that takes a single keys
and a pair of `keys_to_fun/keys_to_optimize`. This results in 12 functions
and comments to maintain.

I removed the overloads and just use a default parameter for
`keys_to_optimize` and grouped the constructors to 3 subsections to reduce
the comments required.

I understand that this has the downside that an empty `keys_to_optimize`
actually means to optimize *all* keys rather than *not* optimize any keys.

However, I believe this is a necessary intermediate step towards removing
the `keys_to_func/keys_to_optimize` pair and switching to a more explicit
type like `FactorKeys` that has the potential to be more compact and
future-proof.

I'm ok with leaving this change open until the full refactoring is done to
avoid any potential confusions. Just opening it for more visibility and
discussions.

Topic: cq_sf_factor_ctors
Reviewers: aaron, bradley, nathan, hayk
GitOrigin-RevId: 6a3ffc591f73ebb1390cae0e7a1ca755389d1e2e