[SymForce] Add metadata field to ResidualBlock
Adds an optional dictionary to the ResidualBlock dataclass for misc.
metadata, as well as a helper function for setting fields in the
metadata. This is useful, e.g., when creating ResidualBlocks from a
TimestepObjective when you want to associate a specific timestep with
the ResidualBlock. For example, in a TimestepSubProblem:

```
my_residual_blocks["my_timestep_objective"] = [
    MyTimestepObjective.residual_at_timestep(
        args,
    ).set_metadata("timestep", k)
    for k in range(self.timesteps)
]
```

Topic: add_metadata_to_residual_blocks
GitOrigin-RevId: 431043bf5f70e3917f27d36206b3282984548b32