[SymForce] Raise exception during codegen if name is not valid
I foolishly tried to generate a function with the name `my_func.h`, and
found that we don't raise any errors and simply generate invalid C++
code. Added a check to codegen that checks the function name is a valid
identifier before the function is generated to avoid this issue.

Topic: symforce_codegen_valid_names
GitOrigin-RevId: 11b74bb07006c9ddc503ee0b955d1fa0f8768234